pyproject.toml
[tool.poetry]
name = "pywikiplain"
version = "0.1.0"
description = ""
authors = ["Dylan Burati <dylanburati@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
maturin = "^0.14.15"
patchelf = "^0.17.2.1"
jupyter = "^1.0.0"
ipython = "^8.11.0"
python-dotenv = "^1.0.0"
scipy = "^1.10.1"
sqlalchemy = "^2.0.5.post1"
arsenal = {git = "https://github.com/timvieira/arsenal.git"}
polars = "^0.16.12"
pyarrow = "^11.0.0"
tqdm = "^4.65.0"
ijson = "^3.2.0.post0"
spacy = "^3.5.1"
transformers = "^4.26.1"
cytoolz = "^0.12.1"
mmh3 = "^3.0.0"
zstandard = "^0.20.0"
seaborn = "^0.12.2"
lxml = "^5.0.1"
tenacity = "^8.2.3"
httpx = {extras = ["socks"], version = "^0.26.0"}
cbor2 = "^5.6.1"
pypocketmap = {git = "https://github.com/dylanburati/pypocketmap.git"}

[tool.poetry.group.dev.dependencies]
mypy = "^1.1.1"
pylint = "^2.17.0"
black = "^23.1.0"

[build-system]
requires = ["maturin>=0.14,<0.15"]
build-backend = "maturin"

[tool.maturin]
features = ["pyo3/extension-module"]

[tool.pylint]
[tool.pylint.MAIN]
extension-pkg-allow-list = "wikiplain"

[tool.pylint.BASIC]
good-names = ['i,j,k,v,x,y,dx,dy,fp']

[tool.pylint.'MESSAGES CONTROL']
disable = ['missing-module-docstring', 'R1710', 'R1705', 'superfluous-parens']